% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_ID}
\alias{check_ID}
\title{Check ID Column Format and Establishment Number Consistency}
\usage{
check_ID(data, ID_column, estab_no_column)
}
\arguments{
\item{data}{A data frame containing the ID and establishment number columns.}

\item{ID_column}{A string specifying the name of the ID column in `data`.}

\item{estab_no_column}{A string specifying the name of the establishment number column in `data`.}
}
\value{
A list containing messages and row indices of any rows with issues.
        If no issues are found, the function returns `NULL`.
}
\description{
This function checks that all IDs in the specified ID column of a data frame:
\itemize{
  \item Have exactly 10 characters in length.
  \item Contain a substring from position 3 to 5 that matches the corresponding value
        in the establishment number column.
}
}
\examples{
data <- data.frame(ID = c('1233344444', '3455666666'),
                   estab_no = c('333', '555'))
check_ID(data, "ID", "estab_no")

}
