% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_person_to_data}
\alias{match_person_to_data}
\title{Title}
\usage{
match_person_to_data(
  FN,
  LN,
  data,
  FN_column = "FN",
  LN_column = "LN",
  UPI_column = "UPI",
  ...
)
}
\arguments{
\item{FN}{First name}

\item{LN}{Last name}

\item{data}{Data frame containing name and UPI columns}

\item{FN_column}{Column name for first names (default: "First Name")}

\item{LN_column}{Column name for surnames (default: "Surname")}

\item{UPI_column}{Column name for Unique Person Identifer (UPI) (default: "UPI")}

\item{...}{parameters passed to clean_name().}
}
\value{
A list with UPI, people row, and a message
}
\description{
Title
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
botanists <- data.frame(
  `First Name` = c("Carl", "José", "Alexander", "Agnes", "  Jane  "),
  Surname = c("Linnæus", "Banks", "Humboldt", "Arber", "Coldstream"),
  UPI = c("CL001", "JB002", "AH003", "AA004", "JC005"),
  stringsAsFactors = FALSE
)

# Basic exact match with accents
match_person_to_data("Jose", "Banks", botanists)  # Matches "José"

# Name with extra spaces
match_person_to_data("Jane", "Coldstream", botanists)  # Matches "  Jane  "

# Lowercase input, uppercase and accent in data
match_person_to_data("carl", "linnaeus", botanists)  # Matches "Carl" "Linnæus"

# Swapped names
match_person_to_data("Linnæus", "Carl", botanists)  # Swap detection

# No match
match_person_to_data("Greg", "Mendel", botanists)  # No match

botanists <- data.frame(
  `First Name` = c("Carl", "José", "Alexander", "Agnes", "  Jane  ", "Jake"),
  Surname = c("Linnæus", "Banks", "Humboldt", "Arber", "Coldstream", "Banks"),
  UPI = c("CL001", "JB002", "AH003", "AA004", "JC005", "CL002"),
  stringsAsFactors = FALSE
)

# LN only with multiple individuals
match_person_to_data("John", "Banks", botanists)  # Matches "José"
\dontshow{\}) # examplesIf}
}
